requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_annotated-doc_1762868213/work
annotated-types @ file:///home/conda/feedstock_root/build_artifacts/annotated-types_1733247046149/work
anyio @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_anyio_1767719033/work
attrs==25.4.0
backports.zstd @ file:///Users/runner/miniforge3/conda-bld/bld/rattler-build_backports.zstd_1767045020/work
blockbuster==1.5.26
Brotli @ file:///Users/runner/miniforge3/conda-bld/brotli-split_1764017116075/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1767500808759/work/certifi
cffi==2.0.0
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1760437218288/work
click @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_click_1764518652/work
cloudpickle==3.1.2
contourpy @ file:///Users/runner/miniforge3/conda-bld/bld/rattler-build_contourpy_1769156091/work
cryptography==46.0.3
cycler @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_cycler_1764466758/work
dataclasses-json==0.6.7
deepgram-sdk==5.3.1
distro==1.9.0
dnspython @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_dnspython_1757292856/work
email-validator @ file:///home/conda/feedstock_root/build_artifacts/email-validator-meta_1756221362582/work
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1763918002538/work
fastapi @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_fastapi-core_1767625503/work
fastapi-cli @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_fastapi-cli_1766435117/work
filetype==1.2.0
fonttools @ file:///Users/runner/miniforge3/conda-bld/fonttools_1765632680641/work
forbiddenfruit==0.1.4
frozenlist==1.8.0
google-auth==2.47.0
google-genai==1.60.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-health-checking==1.76.0
grpcio-tools==1.75.1
h11 @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_h11_1767729720/work
h2 @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_h2_1756364871/work
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
httpcore @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_httpcore_1745602916/work
httptools @ file:///Users/runner/miniforge3/conda-bld/httptools_1762504070708/work
httpx @ file:///home/conda/feedstock_root/build_artifacts/httpx_1733663348460/work
httpx-sse==0.4.3
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1760286409563/work
importlib_metadata==8.7.1
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_jinja2_1764517220/work
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema_rs==0.29.1
kiwisolver @ file:///Users/runner/miniforge3/conda-bld/bld/rattler-build_kiwisolver_1762488860/work
langchain==1.2.7
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.7
langchain-google-genai==4.2.0
langchain-openai==1.1.7
langchain-text-splitters==1.1.0
langgraph==1.0.7
langgraph-api==0.7.9
langgraph-checkpoint==4.0.0
langgraph-cli==0.4.12
langgraph-prebuilt==1.0.7
langgraph-runtime-inmem==0.23.1
langgraph-sdk==0.3.3
langsmith==0.6.4
markdown-it-py @ file:///home/conda/feedstock_root/build_artifacts/markdown-it-py_1754951200865/work
MarkupSafe @ file:///Users/runner/miniforge3/conda-bld/markupsafe_1759055254345/work
marshmallow==3.26.2
matplotlib==3.10.8
mdurl @ file:///home/conda/feedstock_root/build_artifacts/mdurl_1733255585584/work
multidict==6.7.0
munkres==1.1.4
mypy_extensions==1.1.0
numpy @ file:///Users/runner/miniforge3/conda-bld/bld/rattler-build_numpy_1768085572/work/dist/numpy-2.4.1-cp311-cp311-macosx_11_0_arm64.whl#sha256=7c20868141cbb6ab1f03ebfa4b51c4dc5adb6e51c0e1f539d0b99c4cd215f70f
openai==2.15.0
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.5
ormsgpack==1.12.2
packaging==25.0
pillow @ file:///Users/runner/miniforge3/conda-bld/bld/rattler-build_pillow_1758208741/work
propcache==0.4.1
protobuf==6.33.4
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pydantic_1764434463/work
pydantic-extra-types @ file:///home/conda/feedstock_root/build_artifacts/pydantic-extra-types_1767220962105/work
pydantic-settings @ file:///home/conda/feedstock_root/build_artifacts/pydantic-settings_1762786248253/work
pydantic_core @ file:///Users/runner/miniforge3/conda-bld/bld/rattler-build_pydantic-core_1762989051/work
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1750615794071/work
PyJWT==2.10.1
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pyparsing_1769003998/work
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1733217236728/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dateutil_1751104122/work
python-dotenv @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dotenv_1761503229/work
python-multipart @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-multipart_1767725812/work
PyYAML @ file:///Users/runner/miniforge3/conda-bld/pyyaml_1758891744044/work
regex==2026.1.15
requests @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_requests_1766926974/work
requests-toolbelt==1.0.0
rich @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_rich_1760026188/work/dist
rich-toolkit @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_rich-toolkit_1765985144/work
rsa==4.9.1
shellingham @ file:///home/conda/feedstock_root/build_artifacts/shellingham_1762858192781/work
six @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_six_1753199211/work
sniffio @ file:///home/conda/feedstock_root/build_artifacts/sniffio_1762941432086/work
SQLAlchemy==2.0.46
sse-starlette==2.1.3
starlette @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_starlette_1762016292/work
structlog==25.5.0
tenacity==9.1.2
tiktoken==0.12.0
tomli @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_tomli_1768146676/work
tornado @ file:///Users/runner/miniforge3/conda-bld/tornado_1765836484236/work
tqdm==4.67.1
truststore==0.10.4
typer==0.21.1
typer-slim==0.21.1
typing-inspect==0.9.0
typing-inspection @ file:///home/conda/feedstock_root/build_artifacts/typing-inspection_1764158251755/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1756220668/work
unicodedata2 @ file:///Users/runner/miniforge3/conda-bld/unicodedata2_1763054903095/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1767817748113/work
uuid_utils==0.14.0
uvicorn @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_uvicorn_1766332899/work
uvloop @ file:///Users/runner/miniforge3/conda-bld/uvloop_1762472760210/work
watchfiles @ file:///Users/runner/miniforge3/conda-bld/watchfiles_1760456839480/work
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
sounddevice
yt-dlp
trafilatura
